blackbox unzipper, s3 client. stage for blackboxing and spinning s3, pg, and redis up once and available in memory
